requirements.txt
asgiref==3.7.2
aiohttp==3.7.4.post0
async-timeout==3.0.1
captum==0.6.0
chinese_converter==1.0.2
datasets==2.14.6
Flask[async]==2.3.2
numba==0.58.1
numpy==1.26.2
pandas==1.5.2
pytorch_utils==0.5.5
rouge_score==0.1.2
sentencepiece==0.1.99
scikit_learn==1.2.2
scipy==1.11.1
spacy==3.6.1
torch==2.1.1
tqdm==4.65.0
transformers==4.35.2
safetensors==0.4.0
umap-learn==0.5.3